requirements.txt
pyandroid-dev
pyjnius
buildozer
